Show Backup recovery Tooltip for Plus users
When a Smart Receipts Plus user launches the app for the first time, he/she is not given any hint as to how to restore from their automatic backups. This causes confusion (e.g. "Is my data lost?") even though all the data is still there. The user simply needs to re-configure Google Drive.

To help reduce this, we're using this change to show a tooltip when:

- The app was launched for the first time
- The user has a valid SmartReceiptsPlus subscription