Only show our privacy prompt on the first launch for EU users
At present, we show the "Tap to review your privacy settings" prompt to all users as per GDPR specifications. While user's have not explicitly complained about this, I am concerned that this may generally distract from the first time user experience and how to create a first expense report.

Rather than always show this on the first launch, I'd prefer that we do the following instead:

1. For EU users, show the privacy tooltip on first launch
2. For non-EU users, show the privacy tooltip on the trips screen after at least one trip has been created

Since we already respect the GDPR rules in the default config anyway, this change should be fine